pr bug/rest-204 review feedback
