+++ api/client.go
@@ -146,7 +146,7 @@ func (c Client) REST(method string, p string, body io.Reader, data interface{}) 
 		return handleHTTPError(resp)
 	}
 
-	if resp.StatusCode == http.StatusNoContent && resp.ContentLength == 0 {
+	if resp.StatusCode == http.StatusNoContent {
 		return nil
 	}
